src/backend/requirements-vector-db.txt
# Vector Database Provider dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
weaviate-client>=3.15.4
pydantic>=1.10.7
sentence-transformers>=2.2.2
numpy>=1.24.3
requests>=2.28.2
